Store all migrations in @ordered_migrations
When checking if a migration is safe or not, it is sometimes useful to
look at the ordering of migrations rather than each line in the
migration in isolation. E.g. if you're dropping an index before creating
one you might experience slow queries during the creation of the new
one, but to notice that you need to know the order of them since each
line is considered safe.

This stores all migrations in an instance variable so that we can check
for dangerous ordering.