Add check for deleting an index before creating the new one
When replacing an index in the database it's easy to not think about
the time between the deleting of the old index and creating the new one,
especially if it's on a large table.
If it's a heavily used table, not having an index will make the database
spend a lot of time running those queries until the new index is in
place. This could slow down your web server by having all web threads
waiting for the queries to finish.

This adds a check to see if you're dropping the old index before you've
created the new one recommends creating the new index before dropping
the old one.