Set Postgres statement and lock timeouts in milliseconds
Postgres expects a number argument for statement and lock timeouts to be
in milliseconds. This results in the recently released migration-specific
timeouts being shorter than intended if the user sets them with
`StrongMigrations.statement_timeout=` and `.lock_timeout=`, as PG
interprets seconds as milliseconds.

This change updates the checker to multiply the result by 1,000 if the
database is PG.

Co-authored-by: Susanna Kosonen <kosonen.susanna@gmail.com>