Experimental support for SQL migration inspection on Postgres
Opening for feedback on the general idea.

This PR is a proof-of-concept for analysing the safety of raw SQL migrations
using [pg_query][1]. In this case, showing the safety of a `NOT VALID` foreign
key addition vs the unsafe option.

Thoughts? One obvious downside is that this is Postgres-specific, so MySQL users
would be out of luck. However, that probably shouldn't be a reason on its own to
not add support for Postgres. I think this could be adapted even to analyse the
SQL generated by the Rails migrations DSL methods too?

[1]: https://github.com/lfittl/pg_query