Check for PostgreSQL index corruption in all Rails environments
Previously, the PostgreSQL index corruption bug was checked only in
production environments.

However, this can cause the issue to go unnoticed up to the point where
a project tries to release to production and migrate the database. CI
systems often run in the Rails test environment causing this check to
silently be skipped. This scenario happened to a project of mine,
requiring quick action to undo a series of steps we thought were being
checked in CI.

Projects can continue to use the StrongMigrations.target_version setting
if their developer or test database version differs from production.